Separate types and utils into discrete files
